Only MIT harvests write metadata to public CDN bucket
Why these changes are being introduced:
It was determined that OGM harvests will not need to write source and normalized metadata to
the MIT public CDN bucket.

How this addresses that need:
* Harvest pipeline step 'write_source_and_normalized' moved to harvester specific method on MITHarvester

Side effects of this change:
* Writing source and normalized metadata to public CDN bucket will only happen for MIT harvests

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109
