+++ harvester/cli.py
@@ -78,6 +78,44 @@ def ping(ctx: click.Context) -> None:
     help="filter for files modified before this date; format YYYY-MM-DD.",
 )
 @click.option(
+    "-o",
+    "--output-file",
+    required=False,
+    type=str,
+    help="Filepath to write single, combined JSONLines file of normalized MITAardvark "
+    "metadata for ALL harvested records.  This is the expected format for the "
+    "TIMDEX pipeline.",
+)
+@click.pass_context
+def harvest(
+    ctx: click.Context,
+    harvest_type: str,
+    from_date: str,
+    until_date: str,
+    output_file: str,
+) -> None:
+    """Harvest command with sub-commands for different sources."""
+    ctx.obj["HARVEST_TYPE"] = harvest_type
+    ctx.obj["FROM_DATE"] = from_date
+    ctx.obj["UNTIL_DATE"] = until_date
+    ctx.obj["OUTPUT_FILE"] = output_file
+
+
+# Attach harvest group to main command
+main.add_command(harvest)
+
+
+@harvest.command(name="mit")
+@click.option(
+    "-i",
+    "--input-files",
+    required=True,
+    envvar="S3_RESTRICTED_CDN_ROOT",
+    type=str,
+    help="Directory location of source record zip files (may be local or s3). Defaults to"
+    " env var S3_RESTRICTED_CDN_ROOT if not set.",
+)
+@click.option(
     "-osd",
     "--output-source-directory",
     required=False,
@@ -96,48 +134,6 @@ def ping(ctx: click.Context) -> None:
     help="Directory to write normalized MITAardvark metadata for EACH harvested record "
     "file with naming convention '<identifier>.aardvark.normalized.json'. Defaults "
     "to env var S3_PUBLIC_CDN_ROOT if not set.",
-)
-@click.option(
-    "-o",
-    "--output-file",
-    required=False,
-    type=str,
-    help="Filepath to write single, combined JSONLines file of normalized MITAardvark "
-    "metadata for ALL harvested records.  This is the expected format for the "
-    "TIMDEX pipeline.",
-)
-@click.pass_context
-def harvest(
-    ctx: click.Context,
-    harvest_type: str,
-    from_date: str,
-    until_date: str,
-    output_source_directory: str,
-    output_normalized_directory: str,
-    output_file: str,
-) -> None:
-    """Harvest command with sub-commands for different sources."""
-    ctx.obj["HARVEST_TYPE"] = harvest_type
-    ctx.obj["FROM_DATE"] = from_date
-    ctx.obj["UNTIL_DATE"] = until_date
-    ctx.obj["OUTPUT_SOURCE_DIRECTORY"] = output_source_directory
-    ctx.obj["OUTPUT_NORMALIZED_DIRECTORY"] = output_normalized_directory
-    ctx.obj["OUTPUT_FILE"] = output_file
-
-
-# Attach harvest group to main command
-main.add_command(harvest)
-
-
-@harvest.command(name="mit")
-@click.option(
-    "-i",
-    "--input-files",
-    required=True,
-    envvar="S3_RESTRICTED_CDN_ROOT",
-    type=str,
-    help="Directory location of source record zip files (may be local or s3). Defaults to"
-    " env var S3_RESTRICTED_CDN_ROOT if not set.",
 )
 @click.option(
     "-s",
@@ -164,6 +160,8 @@ main.add_command(harvest)
 def harvest_mit(
     ctx: click.Context,
     input_files: str,
+    output_source_directory: str,
+    output_normalized_directory: str,
     sqs_topic_name: str,
     preserve_sqs_messages: bool,
     skip_eventbridge_events: bool,
@@ -177,8 +175,8 @@ def harvest_mit(
         sqs_topic_name=sqs_topic_name,
         preserve_sqs_messages=preserve_sqs_messages,
         skip_eventbridge_events=skip_eventbridge_events,
-        output_source_directory=ctx.obj["OUTPUT_SOURCE_DIRECTORY"],
-        output_normalized_directory=ctx.obj["OUTPUT_NORMALIZED_DIRECTORY"],
+        output_source_directory=output_source_directory,
+        output_normalized_directory=output_normalized_directory,
         output_file=ctx.obj["OUTPUT_FILE"],
     )
     results = harvester.harvest()
@@ -223,8 +221,6 @@ def harvest_ogm(
         from_date=ctx.obj["FROM_DATE"],
         include_repositories=include_list,
         exclude_repositories=exclude_list,
-        output_source_directory=ctx.obj["OUTPUT_SOURCE_DIRECTORY"],
-        output_normalized_directory=ctx.obj["OUTPUT_NORMALIZED_DIRECTORY"],
         output_file=ctx.obj["OUTPUT_FILE"],
     )
